infra/requirements.txt
aws-cdk-lib>=2.134.0
constructs>=10.0.0
